Add assignment dialogs

+++ website/src/main/kotlin/de/westermann/robots/website/toolkit/view/View.kt
@@ -26,13 +26,13 @@ abstract class View(
             element.style.display = if (value) "block" else "none"
         }
 
-    fun offsetLeft(element: HTMLElement): Double = element.offsetLeft + ((element.parentElement as? HTMLElement)?.let {
+    fun offsetLeft(element: HTMLElement): Int = element.offsetLeft + ((element.parentElement as? HTMLElement)?.let {
         offsetLeft(it)
-    } ?: 0.0)
+    } ?: 0)
 
-    fun offsetTop(element: HTMLElement): Double = element.offsetTop + ((element.parentElement as? HTMLElement)?.let {
+    fun offsetTop(element: HTMLElement): Int = element.offsetTop + ((element.parentElement as? HTMLElement)?.let {
         offsetTop(it)
-    } ?: 0.0)
+    } ?: 0)
 
     init {
         init()
